pyproject.toml
[tool.poetry]
name = "kvdb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-datastore = "^2.19.0"
pydantic = "^2.8.2"
google-cloud-firestore = "^2.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.4.2"
pyright = "^1.1.373"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
